backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
psycopg2-binary
PyJWT
python-dotenv
google-auth
requests       # <-- add this
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
PyJWT==2.9.0
google-auth==2.35.0
